refactor: deepen catalog surface write policy
(cherry picked from commit 5e148566c541e8f786795e067e844dade05331c7)
